Refactored file upload functions for downloads
Added attachment flags to file upload functions so that the file downloads
instead of opens in a new window. Uploaded file also configured to
use display name as default value when downloaded. Cloudinary default
was 'file'.
Created a helper function to take the display name and double escape
all periods in the download url.